build(vcpkg): 第三方库切换 vcpkg 并完成迁移收尾，单测 21 模块全绿
- jsoncpp/openssl 等三方库改经 vcpkg manifest（vcpkg.json +
  x64-linux-bw 自定义 triplet）提供，删除 in-tree 源树；
  mak/CMake 构建脚本对接 staging 产物（详见 docs/vcpkg-migration.md）
- gcc15/C++20/OpenSSL3/MySQL8/Python3.13 移植修复（§4.1 共 10 项，
  含 watcher int64 截断、ECDSA DER 签名长度、snapshot_helper
  逗号表达式等真实缺陷）并补对应单测
- 运行期初始化修复（docs/python-313-migration.md §3.1）：补源码包
  缺失的 bw_site.py（随 pyscript 跟踪 + third_party_python.mak
  安装规则）；bw_platform_info.cpp 增加 Debian→el7 回退对齐构建侧，
  修复 lib-dynload-unknown 目录名不匹配导致 C 扩展 stdlib
  （_struct 等）全部无法导入、Script::init 失败的级联
- 全量单测 21 模块 856 用例无失败（make bw-run-all-unit-tests）

+++ programming/bigworld/lib/entitydef/unit_test/test_script_data_sink.cpp
@@ -404,11 +404,13 @@ TEST_F( EntityDefUnitTestHarness, Script_ScriptDataSink_wstring )
 	RETURN_ON_FAIL_CHECK( PyUnicode_Check( resultObject.get() ) );
 	BW::wstring resultStr;
 	CHECK_EQUAL( 0, Script::setData( resultObject.get(), resultStr, "" ) );
-	// No CHECK_EQUAL for BW::wstring
+	// No CHECK_EQUAL for BW::wstring; compare the code points instead.
+	// C++20 deleted operator<<( ostream&, wchar_t ), so the elements have to
+	// be widened to a streamable integer type.
 	CHECK_EQUAL( input.size(), resultStr.size() );
-	CHECK_EQUAL( input[ 0 ], resultStr[ 0 ] );
-	CHECK_EQUAL( input[ 1 ], resultStr[ 1 ] );
-	CHECK_EQUAL( input[ 2 ], resultStr[ 2 ] );
+	CHECK_EQUAL( (uint32)input[ 0 ], (uint32)resultStr[ 0 ] );
+	CHECK_EQUAL( (uint32)input[ 1 ], (uint32)resultStr[ 1 ] );
+	CHECK_EQUAL( (uint32)input[ 2 ], (uint32)resultStr[ 2 ] );
 }
